fix issue #94 separate concerns with new component [2:12pm]
Problem:
i.e. 5 reminder bars to click on. The last click saved the data.
But it would save it 5x yielding a "Fatal: Too many Clients" error from prisma & psql

the reminder bars are dynamic UI and attempted saving based on index of Reminder.

created redux state that toggles reminder bars from visibility.
<Reminder> -> <UpdateThumb> {saveData} -> back to <Reminder> ...

no pun but also comes with a reminder of showing the <Statistics.png> image which is in the navbar to go see the saved data.